requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.3
django-appconf==1.0.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dnspython==2.7.0
drf-extra-fields==3.7.0
executing==2.1.0
filetype==1.2.0
gunicorn==20.1.0
idna==3.10
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
